refactor(mespapiers): Extract Search block from Home/Content
no code change
