+++ packages/cozy-mespapiers-lib/src/components/Home/Content.jsx
@@ -1,15 +1,10 @@
-import React, { useMemo, useState } from 'react'
+import React, { useMemo } from 'react'
 import PropTypes from 'prop-types'
 import uniqBy from 'lodash/uniqBy'
 
-import { models } from 'cozy-client'
 import Empty from 'cozy-ui/transpiled/react/Empty'
-import Box from 'cozy-ui/transpiled/react/Box'
 import { useI18n } from 'cozy-ui/transpiled/react/I18n'
-import useBreakpoints from 'cozy-ui/transpiled/react/hooks/useBreakpoints'
 
-import ThemesFilter from '../ThemesFilter'
-import SearchInput from '../SearchInput'
 import { buildFilesWithContacts } from '../Papers/helpers'
 import PaperGroup from '../Papers/PaperGroup'
 import FeaturedPlaceholdersList from '../Placeholders/FeaturedPlaceholdersList'
@@ -19,14 +14,9 @@ import { useMultiSelection } from '../Hooks/useMultiSelection'
 import { getFeaturedPlaceholders } from '../../helpers/findPlaceholders'
 import HomeCloud from '../../assets/icons/HomeCloud.svg'
 import SearchResult from '../SearchResult/SearchResult'
-
-import { filterPapersByThemeAndSearchValue } from '../Home/helpers'
-import HomeToolbar from '../Home/HomeToolbar'
-import FilterButton from '../Home/FilterButton'
-
-const {
-  themes: { themesList }
-} = models.document
+import { filterPapersByThemeAndSearchValue } from './helpers'
+import HomeToolbar from './HomeToolbar'
+import SearchHeader from './SearchHeader'
 
 const Content = ({
   contacts,
@@ -38,22 +28,10 @@ const Content = ({
   searchValue,
   setSearchValue
 }) => {
-  const { isMobile } = useBreakpoints()
   const { t } = useI18n()
   const scannerT = useScannerI18n()
   const { isMultiSelectionActive } = useMultiSelection()
   const { papersDefinitions } = usePapersDefinitions()
-  const [isThemesFilterDisplayed, setIsThemesFilterDisplayed] = useState(
-    !isMultiSelectionActive
-  )
-
-  const handleThemesFilterDisplayed = isDisplayed => {
-    setIsThemesFilterDisplayed(isDisplayed)
-  }
-
-  const handleThemeSelection = nextValue => {
-    setSelectedTheme(oldValue => (nextValue === oldValue ? '' : nextValue))
-  }
 
   const allPapersByCategories = useMemo(
     () =>
@@ -91,35 +69,12 @@ const Content = ({
   return (
     <>
       {isMultiSelectionActive && <HomeToolbar />}
-
-      <div className="u-flex u-flex-column-s u-mv-1 u-ph-half">
-        <Box className="u-flex u-flex-items-center u-mb-half-s" flex="1 1 auto">
-          <SearchInput
-            value={searchValue}
-            onChange={ev => setSearchValue(ev.target.value)}
-            onFocus={() => handleThemesFilterDisplayed(false)}
-          />
-          {(!isThemesFilterDisplayed || isMobile) && (
-            <FilterButton
-              badge={{
-                active: !isThemesFilterDisplayed,
-                content: selectedTheme ? 1 : 0
-              }}
-              onClick={() => handleThemesFilterDisplayed(prev => !prev)}
-            />
-          )}
-        </Box>
-        <Box className="u-flex u-flex-justify-center" flexWrap="wrap">
-          {isThemesFilterDisplayed && (
-            <ThemesFilter
-              items={themesList}
-              selectedTheme={selectedTheme}
-              handleThemeSelection={handleThemeSelection}
-            />
-          )}
-        </Box>
-      </div>
-
+      <SearchHeader
+        selectedTheme={selectedTheme}
+        setSelectedTheme={setSelectedTheme}
+        searchValue={searchValue}
+        setSearchValue={setSearchValue}
+      />
       {allPapersByCategories.length > 0 ? (
         !isSearching ? (
           <PaperGroup
@@ -138,7 +93,6 @@ const Content = ({
           className="u-ph-1"
         />
       )}
-
       {!isMultiSelectionActive && (
         <FeaturedPlaceholdersList featuredPlaceholders={featuredPlaceholders} />
       )}
